+++ R/plot.R
@@ -1,3 +1,467 @@
+#' @title Plot Coloured Score Table
+#'
+#' @description
+#' Plots a coloured table of summarised scores obtained using
+#' \code{\link{eval_forecasts}}
+#'
+#' @param summarised_scores A data.frame of summarised scores as produced by
+#' \code{\link{eval_forecasts}}
+#' @param select_metrics A character vector with the metrics to show. If set to
+#' \code{NULL} (default), all metrics present in \code{summarised_scores} will
+#' be shown
+#' @return A ggplot2 object with a coloured table of summarised scores
+#' @importFrom ggplot2 ggplot aes element_blank element_text labs coord_cartesian
+#' @importFrom data.table setDT melt
+#' @importFrom stats sd
+#' @export
+#'
+#' @examples
+#' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_wide,
+#'                                        by = c("model", "id", "horizon"),
+#'                                        summarise_by = "model")
+#' scoringutils::score_table(scores)
+
+score_table <- function(summarised_scores,
+                           select_metrics = NULL) {
+
+
+  # identify metrics -----------------------------------------------------------
+  # identify metrics by looking at which of the available column names
+  # are metrics. All other variables are treated as identifier variables
+
+  all_metrics <- list_of_avail_metrics()
+
+  metrics <- names(summarised_scores)[names(summarised_scores) %in% all_metrics]
+  id_vars <- names(summarised_scores)[!(names(summarised_scores) %in% all_metrics)]
+
+
+  # compute scaled values ------------------------------------------------------
+  # scaling is done in order to colour the different scores
+  # for most metrics larger is worse, but others like bias are better if they
+  # are close to zero and deviations in both directions is worse
+  data.table::setDT(summarised_scores)
+
+  # define which metrics are scaled using min (larger is worse) and
+  # which not (metrics like bias where deviations in both directions are bad)
+  metrics_non_min <- c("bias", "coverage_deviation")
+  metrics_no_color <- c("calibration")
+  metrics_min <- setdiff(metrics, c(metrics_non_min, metrics_no_color))
+
+  if (!is.null(select_metrics)) {
+    metrics_non_min <- metrics_non_min[metrics_non_min %in% metrics_select]
+    metrics_min <- metrics_min[metrics_min %in% metrics_select]
+  }
+
+  # write scale functions that can be used in data.table
+  scale <- function(x) {
+    scaled <- x / sd(x, na.rm = TRUE)
+    return(scaled)
+  }
+  scale_min <- function(x) {
+    scaled <- (x - min(x)) / sd(x, na.rm = TRUE)
+    return(scaled)
+  }
+
+  # pivot longer and add scaled values
+  df <- data.table::melt(summarised_scores, value.vars = metrics,
+                         id.vars = id_vars,
+                         variable.name = "metric")
+
+  df[metric %in% metrics_min, value_scaled := scale_min(value),
+     by = metric]
+  df[metric %in% metrics_non_min, value_scaled := scale(value),
+     by = metric]
+  df[metric %in% metrics_no_color, value_scaled := 0,
+  by = metric]
+
+  # create an identifier column by concatinating all columns that
+  # are not a metric
+  identifier_columns <- names(df)[!names(df) %in%
+                                    c("metric", "value", "value_scaled")]
+  df[, identif := do.call(paste, c(.SD, sep = "_")),
+     .SDcols = identifier_columns]
+
+  # make plot with all metrics that are not NA
+  plot <- ggplot2::ggplot(df[!is.na(value), ],
+                          ggplot2::aes(y = identif, x = metric)) +
+    #ggplot2::geom_tile(fill = "blue") +
+    ggplot2::geom_tile(ggplot2::aes(fill = value_scaled), colour = "white") +
+    ggplot2::geom_text(ggplot2::aes(y = identif, label = round(value, 2))) +
+    ggplot2::scale_fill_gradient2(low = "steelblue", high = "salmon") +
+    ggplot2::theme_light() +
+    ggplot2::theme(legend.title = ggplot2::element_blank(),
+                   legend.position = "none",
+                   axis.text.x = ggplot2::element_text(angle = 45, vjust = 1,
+                                                       hjust=1)) +
+    ggplot2::labs(x = "", y = "") +
+    ggplot2::coord_cartesian(expand=FALSE)
+
+  # colouring for pit_p_val is not ideal
+
+  return(plot)
+
+}
+
+
+
+
+#' @title Plot Correlation Between Metrics
+#'
+#' @description
+#' Plots a coloured table of summarised scores obtained using
+#' \code{\link{eval_forecasts}}
+#'
+#' @param scores A data.frame of scores as produced by
+#' \code{\link{eval_forecasts}}
+#' @param select_metrics A character vector with the metrics to show. If set to
+#' \code{NULL} (default), all metrics present in \code{summarised_scores} will
+#' be shown
+#' @return A ggplot2 object showing a coloured matrix of correlations
+#' between metrics
+#' @importFrom ggplot2 ggplot geom_tile geom_text aes scale_fill_gradient2
+#' element_text labs coord_cartesian theme theme_minimal
+#' @importFrom stats cor na.omit
+#' @importFrom data.table setDT melt
+#' @importFrom forcats fct_relevel fct_rev
+#' @export
+#'
+#' @examples
+#' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_wide,
+#'                                        by = c("model", "id", "horizon"),
+#'                                        summarise_by = c("model", "id"))
+#' scoringutils::correlation_plot(scores)
+
+
+correlation_plot <- function(scores,
+                             select_metrics = NULL) {
+
+  # define possible metrics
+  all_metrics <- list_of_avail_metrics()
+
+
+  # find metrics present
+  metrics <- names(scores)[names(scores) %in% all_metrics]
+
+  # restrict to selected metrics
+  if (!is.null(select_metrics)) {
+    metrics <- metrics[metrics %in% metrics_select]
+  }
+
+  # remove all non metrics and non-numeric columns
+  df <- scores[, .SD, .SDcols = sapply(scores,
+                                 function(x) {
+                                   (all(is.numeric(x))) && all(is.finite(x))
+                                 })]
+  df <- df[, .SD, .SDcols = names(df) %in% metrics]
+
+  # define correlation matrix
+  cor_mat <- round(cor(as.matrix(df)), 2)
+
+  # define function to obtain upper triangle of matrix
+  get_lower_tri <- function(cormat){
+    cormat[lower.tri(cormat)] <- NA
+    return(cormat)
+  }
+
+  # get plot data.frame
+  plot_df <- data.table::setDT(as.data.frame(get_lower_tri(cor_mat)),
+                               keep.rownames = TRUE)[, metric := rn][, rn := NULL]
+  plot_df <- na.omit(data.table::melt(plot_df, id.vars = "metric"))
+
+  # refactor levels according to the metrics
+  metrics <- unique(plot_df$metric)
+  plot_df[, metric := forcats::fct_relevel(metric, metrics)]
+  plot_df[, variable := forcats::fct_relevel(variable, metrics)]
+  plot_df[, variable := forcats::fct_rev(variable)]
+
+  plot <- ggplot2::ggplot(plot_df, ggplot2::aes(x = variable, y = metric,
+                                                fill = value)) +
+    ggplot2::geom_tile(color = "white") +
+    ggplot2::geom_text(ggplot2::aes(y = metric, label = value)) +
+    ggplot2::scale_fill_gradient2(low = "steelblue", mid = "white",
+                                  high = "salmon",
+                                  name = "Correlation",
+                                  breaks = c(-1, -0.5, 0, 0.5, 1)) +
+    ggplot2::theme_minimal() +
+    ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45, vjust = 1,
+                                                       hjust=1)) +
+    ggplot2::labs(x = "", y = "") +
+    ggplot2::coord_cartesian(expand = FALSE)
+
+  return(plot)
+
+}
+
+
+
+
+
+
+#' @title Plot Contributions to the Weighted Interval Score
+#'
+#' @description
+#' Visualise the components of the weighted interval score: penalties for
+#' over-predicition, under-prediction and for a lack of sharpness
+#'
+#' @param scores A data.frame of scores based on quantile forecasts as
+#' produced by \code{\link{eval_forecasts}}
+#' @param x The variable from the scores you want to show on the x-Axis.
+#' Usually this will be "model"
+#' @param group Choose a grouping variable for the plot that gets directly
+#' passed down to ggplot. Default is \code{NULL}
+#' @param relative_contributions show relative contributions instead of absolute
+#' contributions. Default is FALSE and this functionality is not available yet.
+#' @param facet_formula facetting formula passed down to ggplot. Default is
+#' \code{NULL}
+#' @param scales scales argument that gets passed down to ggplot. Only necessary
+#' if you make use of facetting. Default is "free_y"
+#' @param nrow nrow argument that gets passed down to ggplot. Specifies the
+#' number of rows to use for \code{facet_wrap} in ggplot
+#' @param xlab Label for the x-axis. Default is the variable name on the x-axis
+#' @param ylab Label for the y-axis. Default is "WIS contributions"
+#' @return A ggplot2 object showing a contributions from the three components of
+#' the weighted interval score
+#' @importFrom ggplot2 ggplot aes_string aes geom_linerange facet_wrap labs
+#' theme theme_minimal unit
+#' @export
+#'
+#' @examples
+#' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_wide,
+#'                                        by = c("model", "id", "horizon"),
+#'                                        summarise_by = c("model", "horizon"))
+#' scoringutils::wis_components(scores, x = "model", facet_formula = ~ horizon)
+#'
+#' @references
+#' Bracher J, Ray E, Gneiting T, Reich, N (2020) Evaluating epidemic forecasts
+#' in an interval format. \url{https://arxiv.org/abs/2005.12881}
+
+
+wis_components <- function(scores,
+                           x,
+                           group = NULL,
+                           relative_contributions = FALSE,
+                           facet_formula = NULL,
+                           scales = "free_y",
+                           nrow = NULL,
+                           xlab = x,
+                           ylab = "WIS contributions") {
+
+    plot <- ggplot2::ggplot(scores, ggplot2::aes_string(x = x, group = group)) +
+    ggplot2::geom_linerange(ggplot2::aes(ymax = is_underprediction,
+                                         ymin = 0, colour = "Underprediction"),
+                            size = 3) +
+    ggplot2::geom_linerange(ggplot2::aes(ymax = is_overprediction + is_underprediction,
+                                         ymin = is_underprediction,
+                                         colour = "Overprediction"),
+                            size = 3)  +
+    ggplot2::geom_linerange(ggplot2::aes(ymax = sharpness + is_overprediction + is_underprediction,
+                                         ymin = is_overprediction + is_underprediction,
+                                         colour = "Sharpness"),
+                            size = 3) +
+    ggplot2::facet_wrap(facet_formula, nrow = nrow,
+                        scales = scales) +
+    ggplot2::labs(x = xlab, y = ylab) +
+    ggplot2::theme_minimal() +
+    ggplot2::theme(panel.spacing = ggplot2::unit(4, "mm"))
+
+    return(plot)
+
+}
+
+# wis_components_rel <- function(scores, x = "model") {
+#
+#   df <- data.table::melt(scores, measure.vars = c("is_underprediction",
+#                                                   "is_overprediction",
+#                                                   "sharpness"))
+#
+#   ggplot2::ggplot(df, ggplot2::aes(x = horizon, y = value, group = group,
+#                                    fill = variable)) +
+#     ggplot2::geom_bar(position = "fill", stat = "identity")
+#
+#
+# }
+
+
+
+
+
+
+
+
+
+
+#' @title Plot Metrics by Range of the Prediction Interval
+#'
+#' @description
+#' Visualise the metrics by range, e.g. if you are interested how different
+#' interval ranges contribute to the overal interval score, or how sharpness
+#' changes by range.
+#'
+#' @param scores A data.frame of scores based on quantile forecasts as
+#' produced by \code{\link{eval_forecasts}}. Note that "range" must be included
+#' in the \code{summarise_by} argument when running \code{eval_forecasts}
+#' @param y The variable from the scores you want to show on the y-Axis.
+#' This could be something like "interval_score" (the default) or "sharpness"
+#' @param x The variable from the scores you want to show on the x-Axis.
+#' Usually this will be "model"
+#' @param colour Charachter vector of length one used to determine a variable
+#' for colouring dots. The Default is "range".
+#' @param xlab Label for the x-axis. Default is the variable name on the x-axis
+#' @param ylab Label for the y-axis. Default is "WIS contributions"
+#' @return A ggplot2 object showing a contributions from the three components of
+#' the weighted interval score
+#' @importFrom ggplot2 ggplot aes_string aes geom_point geom_line
+#' expand_limits theme theme_light element_text scale_color_continuous labs
+#' @export
+#'
+#' @examples
+#' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_long,
+#'                                        by = c("model", "id", "horizon"),
+#'                                        summarise_by = c("model", "horizon", "range"))
+#' scoringutils::range_plot(scores, x = "model")
+#'
+#' scoringutils::range_plot(scores, y = "sharpness", x = "model")
+
+
+range_plot <- function(scores,
+                       y = "interval_score",
+                       x = "model",
+                       colour = "range",
+                       xlab = x,
+                       ylab = y) {
+
+  ggplot2::ggplot(scores,
+                  ggplot2::aes_string(x = x,
+                                      y = y,
+                                      colour = colour)) +
+    ggplot2::geom_point(size = 2) +
+    ggplot2::geom_line(ggplot2::aes(group = range),
+                       colour = "black",
+                       size = 0.01) +
+    ggplot2::theme_light() +
+    ggplot2::expand_limits(y = 0) +
+    ggplot2::scale_color_continuous(low = "steelblue", high = "salmon") +
+    ggplot2::theme(legend.position = "right",
+                   axis.text.x = ggplot2::element_text(angle = 45, vjust = 1,
+                                                       hjust=1)) +
+    ggplot2::labs(y = ylab,
+                  x = xlab)
+
+}
+
+
+
+
+
+
+
+
+
+
+#' @title Create a Heatmap of a Scoring Metric
+#'
+#' @description
+#' This function can be used to create a heatmap of one metric across different
+#' groups, e.g. the interval score obtained by several forecasting models in
+#' different locations.
+#'
+#' @param scores A data.frame of scores based on quantile forecasts as
+#' produced by \code{\link{eval_forecasts}}.
+#' @param y The variable from the scores you want to show on the y-Axis. The
+#' default for this is "model"
+#' @param x The variable from the scores you want to show on the x-Axis. This
+#' could be something like "horizon", or "location"
+#' @param metric the metric that determines the value and colour shown in the
+#' tiles of the heatmap
+#' @param xlab Label for the x-axis. Default is the variable name on the x-axis
+#' @param ylab Label for the y-axis. Default is "WIS contributions"
+#' @return A ggplot2 object showing a heatmap of the desired metric
+#' @importFrom data.table setDT `:=`
+#' @importFrom ggplot2 ggplot aes_string aes geom_tile geom_text
+#' scale_fill_gradient2 labs element_text coord_cartesian
+#' @export
+#'
+#' @examples
+#' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_long,
+#'                                        by = c("model", "id", "horizon"),
+#'                                        summarise_by = c("model", "horizon"))
+#'
+#' scoringutils::score_heatmap(scores, x = "horizon", metric = "bias")
+#'
+
+
+
+score_heatmap <- function(scores,
+                         y = "model",
+                         x,
+                         metric,
+                         ylab = y,
+                         xlab = x) {
+
+
+  data.table::setDT(scores)
+
+  scores[, eval(metric) := round(get(metric), 2)]
+
+  # scores <- scores %>%
+  #   dplyr::arrange(state, model, interval_score) %>%
+  #   dplyr::group_by(state) %>%
+  #   dplyr::mutate(cov_scaled = (coverage_deviation) / sd(coverage_deviation)) %>%
+  #   dplyr::ungroup() %>%
+  #   dplyr::mutate(model = forcats::fct_reorder(model,
+  #                                              interval_score,
+  #                                              .fun='mean',
+  #                                              .desc = TRUE),
+  #                 state = forcats::fct_reorder(state,
+  #                                              coverage_deviation,
+  #                                              .fun='mean',
+  #                                              .desc = TRUE))
+
+
+  ggplot2::ggplot(scores,
+                  ggplot2::aes_string(y = y,
+                                      x = x,
+                                      fill = metric)) +
+    ggplot2::geom_tile() +
+    ggplot2::geom_text(ggplot2::aes_string(label = metric)) +
+    ggplot2::scale_fill_gradient2(low = "skyblue", high = "red") +
+    ggplot2::labs(y = ylab, x = xlab) +
+    ggplot2::theme(axis.text.x = ggplot2::element_text(angle = 45, vjust = 1,
+                                                       hjust=1)) +
+    ggplot2::coord_cartesian(expand = FALSE)
+
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+## make fixed effects regression and plot random effects?
+
+
+
+
+
+
+
+
+
+
+
 #' #' @title Plot Scores
 #' #'
 #' #' @description
@@ -9,6 +473,7 @@
 #' #' @param facet_formula formula for facetting in ggplot
 #' #' @return list of plots
 #' #' @importFrom ggplot2 ggplot
+#'
 #' #' @examples
 #' #' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_long,
 #' #'                                        summarised = FALSE)
@@ -206,3 +671,73 @@
 #'
 #'
 #'
+#'
+
+
+
+#' #' @title Plot Predictions vs True Values
+#' #'
+#' #' @description
+#' #' Simple plot of true vs observed values
+#' #'
+#' #' @param scores Unsummarised scores as produced by \code{\link{eval_forecasts}}
+#' #' @param by character vector to do grouping by
+#' #' @param facet_formula formula for facetting in ggplot
+#' #' @return list of plots
+#' #' @importFrom ggplot2 ggplot
+#' #'
+#' #' @examples
+#' #' scores <- scoringutils::eval_forecasts(scoringutils::quantile_example_data_long,
+#' #'                                        summarised = FALSE)
+#' #' plot_predictions(d)
+#'
+#'
+#' d <- scoringutils::continuous_example_data
+#' plot_predictions(d)
+#'
+#' d <- scoringutils::quantile_example_data_long
+#'
+#' plot_predictions_quantile <- function(data,
+#'                                       facet_formula = NULL,
+#'                                       facet_or_plot_grid = "facet",
+#'                                       scales = "free_y",
+#'                                       interval_ranges = c(0, 90)) {
+#'
+#'
+#'   # compute quantiles
+#'
+#'   plot <- data %>%
+#'     ggplot2::ggplot(ggplot2::aes(x = id)) +
+#'     ggplot2::geom_point(ggplot2::aes(y = true_values), size = 1) +
+#'     ggplot2::geom_line(ggplot2::aes(y = true_values), lwd = 0.2)
+#'
+#'   for (range in interval_ranges) {
+#'     # lower <- 1 - (interval_ranges / 100) / 2
+#'     # upper <- 1 - lower
+#'
+#'     # check whether we have continuous, binary or quantile forecasts
+#'
+#'     plot <- plot +
+#'       ggplot2::geom_ribbon()
+#'   }
+#'
+#'   if (facet_or_plot_grid == "facet") {
+#'     plot <- plot +
+#'       ggplot2::facet_wrap(facet_formula, scales = scales)
+#'   } else {
+#'     plot <- plot +
+#'       ggplot2::facet_wrap(facet_formula, scales = scales)
+#'   }
+#'
+#'   plot <- plot +
+#'     ggplot2::expand_limits(y = 0) +
+#'     ggplot2::labs(x = "Week", y = "Weekly incident deaths",
+#'                   caption = NULL,
+#'                   col = "Model", fill = "Model") +
+#'     cowplot::theme_cowplot() +
+#'     ggplot2::theme(legend.position = "bottom")
+#'
+#'   return(plot)
+#' }
+#'
+#' plot_predictions_quantile(d)