+++ R/utils.R
@@ -6,18 +6,37 @@ globalVariables(c(".",
                   "coverage_deviation",
                   "CRPS",
                   "DSS",
+                  "identif",
                   "Interval_Score",
+                  "is_overprediction",
+                  "is_underprediction",
                   "LogS",
                   "calibration",
                   "hist",
                   "id",
                   "log_score",
                   "lower",
+                  "metric",
+                  "metrics_select",
                   "model",
                   "pit_p_val",
                   "predictions",
                   "quantile",
+                  "rn",
                   "true_values",
                   "type",
                   "upper",
+                  "value",
+                  "value_scaled",
+                  "variable",
                   "x"))
+
+
+list_of_avail_metrics <- function() {
+  available_metrics <- c("log_score", "sharpness", "bias", "dss", "crps",
+                         "calibration", "coverage_deviation",
+                         "pit_p_val", "pit_sd","interval_score",
+                         "is_underprediction", "is_overprediction")
+
+  return(available_metrics)
+}